\name{abodoutlier}
\alias{abodoutlier}
\alias{abodoutlier}
\docType{package}
\title{
Angle-based outlier detection
}
\description{
Performs angle-based outlier detection on data. A complete, a randomized and a knn based methods are available.
\details{
\tabular{ll}{
Package: \tab abodoutlier\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-08-30\cr
License: \tab MIT License\cr
}

}
\author{
José Jiménez\cr
Maintainer: José Jiménez <jose@jimenezluna.com>
}